requirements.txt
ipython==2.3.0
click==3.3
Pillow==2.6.1
dateutils==0.6.6
